rename GraphicController into Drawer
