+++ code/core/src/controller/LevelController.java
@@ -1,9 +1,10 @@
 package controller;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import graphic.Drawer;
 
 public class LevelController {
-    public LevelController(SpriteBatch batch, GraphicController graphicController) {}
+    public LevelController(SpriteBatch batch, Drawer drawer) {}
 
     public void update() {}
 }
